+++ pyo3-macros-backend/src/pymethod.rs
@@ -93,7 +93,7 @@ pub fn impl_wrap_cfunction_with_keywords(
     let body = impl_call(cls, &spec);
     let slf = self_ty.receiver(cls);
     let py = syn::Ident::new("_py", Span::call_site());
-    let body = impl_arg_params(&spec, Some(cls), body, &py)?;
+    let body = impl_arg_params(&spec, Some(cls), body, &py, false)?;
     let deprecations = &spec.deprecations;
     Ok(quote! {{
         unsafe extern "C" fn __wrap(
@@ -106,6 +106,42 @@ pub fn impl_wrap_cfunction_with_keywords(
                 #slf
                 let _args = #py.from_borrowed_ptr::<pyo3::types::PyTuple>(_args);
                 let _kwargs: Option<&pyo3::types::PyDict> = #py.from_borrowed_ptr_or_opt(_kwargs);
+
+                #body
+            })
+        }
+        __wrap
+    }})
+}
+
+/// Generate function wrapper for PyCFunctionFastWithKeywords
+pub fn impl_wrap_fastcall_cfunction_with_keywords(
+    cls: &syn::Type,
+    spec: &FnSpec<'_>,
+    self_ty: &SelfType,
+) -> Result<TokenStream> {
+    let body = impl_call(cls, &spec);
+    let slf = self_ty.receiver(cls);
+    let py = syn::Ident::new("_py", Span::call_site());
+    let body = impl_arg_params(&spec, Some(cls), body, &py, true)?;
+    Ok(quote! {{
+        unsafe extern "C" fn __wrap(
+            _slf: *mut pyo3::ffi::PyObject,
+            _args: *const *mut pyo3::ffi::PyObject,
+            _nargs: pyo3::ffi::Py_ssize_t,
+            _kwnames: *mut pyo3::ffi::PyObject) -> *mut pyo3::ffi::PyObject
+        {
+            pyo3::callback::handle_panic(|#py| {
+                #slf
+                let _kwnames: Option<&pyo3::types::PyTuple> = #py.from_borrowed_ptr_or_opt(_kwnames);
+                // Safety: &PyAny has the same memory layout as `*mut ffi::PyObject`
+                let _args = _args as *const &pyo3::PyAny;
+                let _kwargs = if let Some(kwnames) = _kwnames {
+                    std::slice::from_raw_parts(_args.offset(_nargs), kwnames.len())
+                } else {
+                    &[]
+                };
+                let _args = std::slice::from_raw_parts(_args, _nargs as usize);
 
                 #body
             })
@@ -142,7 +178,7 @@ pub fn impl_wrap_new(cls: &syn::Type, spec: &FnSpec<'_>) -> Result<TokenStream> 
     let names: Vec<syn::Ident> = get_arg_names(&spec);
     let cb = quote! { #cls::#name(#(#names),*) };
     let py = syn::Ident::new("_py", Span::call_site());
-    let body = impl_arg_params(spec, Some(cls), cb, &py)?;
+    let body = impl_arg_params(spec, Some(cls), cb, &py, false)?;
     let deprecations = &spec.deprecations;
     Ok(quote! {{
         #[allow(unused_mut)]
@@ -172,7 +208,7 @@ pub fn impl_wrap_class(cls: &syn::Type, spec: &FnSpec<'_>) -> Result<TokenStream
     let names: Vec<syn::Ident> = get_arg_names(&spec);
     let cb = quote! { pyo3::callback::convert(_py, #cls::#name(&_cls, #(#names),*)) };
     let py = syn::Ident::new("_py", Span::call_site());
-    let body = impl_arg_params(spec, Some(cls), cb, &py)?;
+    let body = impl_arg_params(spec, Some(cls), cb, &py, false)?;
     let deprecations = &spec.deprecations;
     Ok(quote! {{
         #[allow(unused_mut)]
@@ -200,7 +236,7 @@ pub fn impl_wrap_static(cls: &syn::Type, spec: &FnSpec<'_>) -> Result<TokenStrea
     let names: Vec<syn::Ident> = get_arg_names(&spec);
     let cb = quote! { pyo3::callback::convert(_py, #cls::#name(#(#names),*)) };
     let py = syn::Ident::new("_py", Span::call_site());
-    let body = impl_arg_params(spec, Some(cls), cb, &py)?;
+    let body = impl_arg_params(spec, Some(cls), cb, &py, false)?;
     let deprecations = &spec.deprecations;
     Ok(quote! {{
         #[allow(unused_mut)]
@@ -379,6 +415,7 @@ pub fn impl_arg_params(
     self_: Option<&syn::Type>,
     body: TokenStream,
     py: &syn::Ident,
+    fastcall: bool,
 ) -> Result<TokenStream> {
     if spec.args.is_empty() {
         return Ok(body);
@@ -428,16 +465,7 @@ pub fn impl_arg_params(
         )?);
     }
 
-    let (mut accept_args, mut accept_kwargs) = (false, false);
-
-    for s in spec.attrs.iter() {
-        use crate::pyfunction::Argument;
-        match s {
-            Argument::VarArgs(_) => accept_args = true,
-            Argument::KeywordArgs(_) => accept_kwargs = true,
-            _ => continue,
-        }
-    }
+    let (accept_args, accept_kwargs) = spec.accept_args_kwargs();
 
     let cls_name = if let Some(cls) = self_ {
         quote! { Some(<#cls as pyo3::type_object::PyTypeInfo>::NAME) }
@@ -445,6 +473,24 @@ pub fn impl_arg_params(
         quote! { None }
     };
     let python_name = &spec.python_name;
+
+    let (args_to_extract, kwargs_to_extract) = if fastcall {
+        // _args is a &[&PyAny], _kwnames is a Option<&PyTuple> containing the
+        // keyword names of the keyword args in _kwargs
+        (
+            // need copied() for &&PyAny -> &PyAny
+            quote! { _args.iter().copied() },
+            quote! { _kwnames.map(|kwnames| {
+                kwnames.as_slice().iter().copied().zip(_kwargs.iter().copied())
+            }) },
+        )
+    } else {
+        // _args is a &PyTuple, _kwargs is an Option<&PyDict>
+        (
+            quote! { _args.iter() },
+            quote! { _kwargs.map(|dict| dict.iter()) },
+        )
+    };
 
     // create array of arguments, and then parse
     Ok(quote! {
@@ -462,7 +508,12 @@ pub fn impl_arg_params(
             };
 
             let mut #args_array = [None; #num_params];
-            let (_args, _kwargs) = DESCRIPTION.extract_arguments(_args, _kwargs, &mut #args_array)?;
+            let (_args, _kwargs) = DESCRIPTION.extract_arguments(
+                #py,
+                #args_to_extract,
+                #kwargs_to_extract,
+                &mut #args_array
+            )?;
 
             #(#param_conversion)*
 
@@ -616,32 +667,36 @@ pub fn impl_py_method_def(
     let add_flags = flags.map(|flags| quote!(.flags(#flags)));
     let python_name = spec.null_terminated_python_name();
     let doc = &spec.doc;
-    if spec.args.is_empty() {
-        let wrapper = impl_wrap_noargs(cls, spec, self_ty);
-        Ok(quote! {
-            pyo3::class::PyMethodDefType::Method({
-                pyo3::class::PyMethodDef::noargs(
-                    #python_name,
-                    pyo3::class::methods::PyCFunction(#wrapper),
-                    #doc
-                )
-                #add_flags
-
-            })
-        })
+    let (methoddef_meth, cfunc_variant) = if spec.args.is_empty() {
+        (quote!(noargs), quote!(PyCFunction))
+    } else if spec.can_use_fastcall() {
+        (
+            quote!(fastcall_cfunction_with_keywords),
+            quote!(PyCFunctionFastWithKeywords),
+        )
     } else {
-        let wrapper = impl_wrap_cfunction_with_keywords(cls, &spec, self_ty)?;
-        Ok(quote! {
-            pyo3::class::PyMethodDefType::Method({
-                pyo3::class::PyMethodDef::cfunction_with_keywords(
-                    #python_name,
-                    pyo3::class::methods::PyCFunctionWithKeywords(#wrapper),
-                    #doc
-                )
-                #add_flags
-            })
+        (
+            quote!(cfunction_with_keywords),
+            quote!(PyCFunctionWithKeywords),
+        )
+    };
+    let wrapper = if spec.args.is_empty() {
+        impl_wrap_noargs(cls, spec, self_ty)
+    } else if spec.can_use_fastcall() {
+        impl_wrap_fastcall_cfunction_with_keywords(cls, &spec, self_ty)?
+    } else {
+        impl_wrap_cfunction_with_keywords(cls, &spec, self_ty)?
+    };
+    Ok(quote! {
+        pyo3::class::PyMethodDefType::Method({
+            pyo3::class::PyMethodDef:: #methoddef_meth (
+                #python_name,
+                pyo3::class::methods:: #cfunc_variant (#wrapper),
+                #doc
+            )
+            #add_flags
         })
-    }
+    })
 }
 
 pub fn impl_py_method_def_new(cls: &syn::Type, spec: &FnSpec) -> Result<TokenStream> {